requirements.txt
connexion==2.14.1
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.29.0
requests==2.31.0
swagger-ui-bundle==0.0.9
